requirements.txt
# Core dependencies - Microsoft Agentic Framework
autogen-agentchat~=0.4
autogen-ext[openai,azure]~=0.4
python-dotenv>=1.0.0

# Azure AI Services
azure-ai-projects>=1.0.0
azure-identity>=1.12.0

# Observability and Telemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Optional but recommended
requests>=2.31.0
pydantic>=2.0.0
aiohttp>=3.9.0
